consolidate file cache workers
