+++ app/workers/scheduled_worker_runner.rb
@@ -34,6 +34,7 @@ class ScheduledWorkerRunner < ScheduledWorker
 
   def self.scheduled_workers
     [UpdateExpiredInvoiceWorker, UpdateCountsWorker, UpdateOrganizationPosKindWorker,
+     FileCacheMaintenanceWorker,
      DeactivateExpiredTheftAlertWorker] + [self]
   end
