+++ spec/workers/file_cache_maintenance_worker_spec.rb
@@ -1,11 +1,19 @@
 require "rails_helper"
 
-RSpec.describe CacheAllStolenWorker, type: :job do
+RSpec.describe FileCacheMaintenanceWorker, type: :job do
+  include_context :scheduled_worker
+  include_examples :scheduled_worker_tests
+
+  it "is the correct queue and frequency" do
+    expect(described_class.sidekiq_options["queue"]).to eq "low_priority"
+    expect(described_class.frequency).to be > 5.hours
+  end
+
   describe "output_stolen" do
     it "creates a stolen cache" do
       FactoryBot.create(:stolen_bike)
       FileCacheMaintainer.redis.expire(FileCacheMaintainer.info_id, 0)
-      CacheAllStolenWorker.new.perform
+      described_class.new.perform
       tsv_record = FileCacheMaintainer.files.last
       expect(tsv_record["filename"]).to match "all_stolen_cache.json"
     end
@@ -14,13 +22,25 @@ RSpec.describe CacheAllStolenWorker, type: :job do
   describe "write_stolen" do
     it "outputs the correct stolen" do
       FactoryBot.create(:stolen_bike, serial_number: "party seri8al")
-      cache_all_stolen_worker = CacheAllStolenWorker.new
+      cache_all_stolen_worker = described_class.new
       cache_all_stolen_worker.write_stolen
       result = JSON.parse(File.read(cache_all_stolen_worker.tmp_path))
       expect(result["bikes"].count).to eq(1)
       expect(result["bikes"][0]["serial"]).to eq("party seri8al")
       bike_v2_serializer_keys = %w(id title serial manufacturer_name frame_model year frame_colors thumb large_img is_stock_img stolen stolen_location date_stolen)
       expect(result["bikes"][0].keys).to eq bike_v2_serializer_keys
+    end
+  end
+
+  describe "removing" do
+    it "removes expired files" do
+      t = (Time.current - 3.days).to_i
+      FileCacheMaintainer.reset_file_info("#{t}_all_stolen_cache.json", t)
+      FileCacheMaintainer.update_file_info("current_stolen_bikes.tsv")
+      expect(FileCacheMaintainer.files.count).to eq 2
+      RemoveExpiredFileCacheWorker.new.perform
+      expect(FileCacheMaintainer.files.count).to eq 1
+      expect(FileCacheMaintainer.files.first["filename"]).to eq "current_stolen_bikes.tsv"
     end
   end
 end
